software/requirements.txt
hidapi==0.14.0
mcp2210-python==1.0.4
numpy==1.26.3
pyqtgraph==0.13.3
PySide6==6.5.2
PySide6-Addons==6.5.2
PySide6-Essentials==6.5.2
shiboken6==6.5.2
